requirements.txt
numpy==1.19.2
pandas==1.1.3
plotly==4.11.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.1
retrying==1.3.3
Shapely==1.7.1
six==1.15.0
